requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==2.1.15
django-allauth==0.44.0
django-braces==1.14.0
django-cors-headers==3.5.0
django-oauth-toolkit==1.3.3
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
idna==2.10
joblib==1.0.0
lxml==4.6.2
mysqlclient==2.0.2
numpy==1.19.3
oauthlib==3.1.0
pandas==1.1.5
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
scikit-learn==0.24.0
scipy==1.5.4
selenium==3.141.0
six==1.15.0
sklearn==0.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.1
sqlparse==0.4.1
threadpoolctl==2.1.0
urllib3==1.26.2
waitress==1.4.4
wget==3.2
